requirements.txt
# Core FastAPI Framework
fastapi==0.117.1
uvicorn==0.35.0
python-multipart==0.0.20
starlette==0.48.0

# Database & ORM
SQLAlchemy==2.0.43
alembic==1.17.0
psycopg2-binary==2.9.11
asyncpg==0.30.0

# Authentication & Security
python-jose==3.5.0
passlib==1.7.4
python-decouple==3.8
pydantic==2.12.3
pydantic-settings==2.10.1
bcrypt==4.0.1
cryptography==46.0.1

# AI & Language Models (Current working versions)
langchain==0.3.27
langchain-core==0.3.75
langchain-community==0.3.29
langchain-google-genai==2.0.10
google-generativeai==0.8.5

# Machine Learning & Embeddings
transformers==4.55.4
torch==2.8.0
sentence-transformers==5.1.1
tokenizers==0.21.4
huggingface-hub==0.34.4

# Document Processing
PyPDF2==3.0.1
PyMuPDF==1.26.5
python-docx==1.2.0
pillow==11.3.0
pdf2image==1.17.0
pytesseract==0.3.13

# Vector Database
chromadb==1.0.20

# Utilities & File Handling
aiofiles==24.1.0
httpx==0.28.1
requests==2.32.5
python-dateutil==2.9.0.post0

# Email Services
fastapi-mail==1.5.2
aiosmtplib==3.0.2

# Payment Integration
stripe==13.0.1
razorpay==2.0.0
twilio==9.8.3

# Development & Testing
pytest==8.4.2
pytest-asyncio==1.2.0

# Background Tasks
celery==5.5.3
redis==6.4.0

# Additional Utilities
numpy==2.2.6
pandas==2.3.2
scipy==1.15.3
scikit-learn==1.6.1
matplotlib==3.10.7
lxml==6.0.2
PyYAML==6.0.2
python-dotenv==1.1.1
tenacity==9.1.2
click==8.2.1
tqdm==4.67.1

# Core Dependencies
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
anyio==4.10.0
sniffio==1.3.1
h11==0.16.0
typing-extensions==4.15.0
typing==3.7.4.3
annotated-types==0.7.0
pydantic-core==2.41.4